+++ wellregistry/wellregistry/__init__.py
@@ -1,5 +1,5 @@
 """
-Well Registry project
+Monitoring Locations Registry project
 
 """
 __version__ = '1.2.0dev'